Separate package management actions from CLI
